moved makeCoinData to own helper
